# Speedup Gsf component merging, return components by const reference