--- a/TrackingTools/GeomPropagators/src/AnalyticalTrajectoryExtrapolatorToLine.cc
+++ b/TrackingTools/GeomPropagators/src/AnalyticalTrajectoryExtrapolatorToLine.cc
@@ -46,7 +46,7 @@ AnalyticalTrajectoryExtrapolatorToLine::extrapolateFullState (const TrajectorySt
   //
   TrajectoryStateOnSurface singleState = 
     extrapolateSingleState(*tsos.freeTrajectoryState(),line);
-  if ( !singleState.isValid() || tsos.components().size()==1 )  return singleState;
+  if ( !singleState.isValid() || tsos.singleState() )  return singleState;
   //
   // propagate multiTsos to plane found above
   //