--- a/TrackingTools/GsfTools/interface/CloseComponentsMerger.icc
+++ b/TrackingTools/GsfTools/interface/CloseComponentsMerger.icc
@@ -11,8 +11,118 @@ CloseComponentsMerger<N>::CloseComponentsMerger (int maxNumberOfComponents,
   theMaxNumberOfComponents(maxNumberOfComponents),
   theDistance(distance->clone()) {}
 
+
+
 template <unsigned int N> MultiGaussianState<N>
 CloseComponentsMerger<N>::merge (const MultiState& mgs) const
+{
+
+  auto oldRes = mergeOld(mgs);
+
+  typedef std::vector<SingleStatePtr>  SingleStateVector;
+
+
+  auto ori = mgs.components();
+  SingleStateVector finalComponents;
+
+  int noComp = ori.size();
+  if (noComp <=theMaxNumberOfComponents) return mgs;
+
+
+  while (true) {
+    SingleStateVector merged; merged.reserve(noComp);
+    
+    int  toMerge[noComp];
+    float weights[noComp];
+    bool active[noComp];
+    for (int i=0; i<noComp; ++i) {
+       toMerge[i]=i;
+       active[i]=true;
+       weights[i]=ori[i]->weight();
+    }
+    auto cmp = [&](int i, int j) { return weights[i]> weights[j];};
+
+    std::make_heap(toMerge,toMerge+noComp,cmp);
+
+    auto minDistToMax = [&]()->int {
+      auto mind = std::numeric_limits<double>::max();
+      int im = 0; 
+      auto topI = toMerge[0];
+      auto const & tc = *ori[topI];
+      active[topI]=false;
+      for (int i=0; i<noComp; ++i) {
+         if (!active[i]) continue;
+         auto dist = (*theDistance)(tc,*ori[i]);
+         if (dist<mind) {
+           mind=dist; im = i;
+         }         
+      }
+      return im;
+    };
+
+    auto nComp = noComp;
+    auto nAct=nComp;
+    auto ntoMerge= noComp;
+    while ( (nAct>0) & (nComp > theMaxNumberOfComponents)) {
+      if (nAct==1) { merged.push_back(ori[toMerge[0]]); nAct=0; break;}
+
+      auto ii = minDistToMax();      
+      SingleStateVector comp;
+      comp.push_back(ori[toMerge[0]]);
+      comp.push_back(ori[ii]);
+      active[ii]=false;
+      while( (ntoMerge>0) & (!active[toMerge[0]])) {std::pop_heap(toMerge,toMerge+ntoMerge,cmp); --ntoMerge;}
+      --nComp;
+      nAct-=2;
+      merged.push_back(MultiGaussianStateCombiner<N>().combine(comp));
+      // assert(ntoMerge>=0);
+      // assert(ntoMerge>=nAct);   
+    }
+
+    // std::cout << theMaxNumberOfComponents << ' ' << noComp << ' ' << nComp << ' ' << nAct << ' ' << ntoMerge << std::endl;
+
+    if (nComp <= theMaxNumberOfComponents) { // end game
+
+        /*
+        {
+          int kk=merged.size();
+          for (int i=0; i<noComp; ++i) { if (active[i]) ++kk;}
+          assert(kk==nComp);
+        }
+        */
+ 
+      MultiGaussianStateAssembler<N> result(mgs);
+
+      for (int i=0; i<noComp; ++i) { if (active[i]) result.addState(ori[i]);}
+      for (auto && s : merged) result.addState(s);
+
+      auto res = result.combinedState();
+      
+      /*
+      assert(res.components().size()==oldRes.components().size());
+      std::cout << res.weight() << ' ' << oldRes.weight() << std::endl;
+      for(auto const & c : res.components()) std::cout << c->weight() << '/';
+      std::cout << std::endl;
+      for(auto const & c : oldRes.components()) std::cout << c->weight() <<	'/';
+      std::cout	<< std::endl;
+      */
+
+      return res;
+    }
+
+    // assert(nAct==0);
+    std::swap(ori,merged); noComp=ori.size();
+  }
+
+  // 
+
+}
+
+
+
+
+template <unsigned int N> MultiGaussianState<N>
+CloseComponentsMerger<N>::mergeOld (const MultiState& mgs) const
 {
   typedef std::vector<SingleStatePtr>  SingleStateVector;
   SingleStateVector unmergedComponents(mgs.components());