--- a/TrackingTools/GsfTools/interface/CloseComponentsMerger.icc
+++ b/TrackingTools/GsfTools/interface/CloseComponentsMerger.icc
@@ -3,6 +3,7 @@
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 
 #include <algorithm>
+#include <queue>
 #include <cfloat>
 
 template <unsigned int N>
@@ -32,26 +33,26 @@ CloseComponentsMerger<N>::merge (const MultiState& mgs) const
   while (true) {
     SingleStateVector merged; merged.reserve(noComp);
     
-    int  toMerge[noComp];
     float weights[noComp];
     bool active[noComp];
     for (int i=0; i<noComp; ++i) {
-       toMerge[i]=i;
        active[i]=true;
        weights[i]=ori[i]->weight();
     }
-    auto cmp = [&](int i, int j) { return weights[i]> weights[j];};
 
-    std::make_heap(toMerge,toMerge+noComp,cmp);
+    auto cmp = [&](int i, int j) { return weights[i] > weights[j];};
+    std::priority_queue<int, std::vector<int>, decltype(cmp)> toMerge(cmp);
+    for (int i=0; i<noComp; ++i) toMerge.push(i);
 
     auto minDistToMax = [&]()->int {
       auto mind = std::numeric_limits<double>::max();
       int im = 0; 
-      auto topI = toMerge[0];
+      auto topI = toMerge.top();
       auto const & tc = *ori[topI];
       active[topI]=false;
       for (int i=0; i<noComp; ++i) {
          if (!active[i]) continue;
+         assert(weights[topI]<=weights[i]);
          auto dist = (*theDistance)(tc,*ori[i]);
          if (dist<mind) {
            mind=dist; im = i;
@@ -62,50 +63,49 @@ CloseComponentsMerger<N>::merge (const MultiState& mgs) const
 
     auto nComp = noComp;
     auto nAct=nComp;
-    auto ntoMerge= noComp;
     while ( (nAct>0) & (nComp > theMaxNumberOfComponents)) {
-      if (nAct==1) { merged.push_back(ori[toMerge[0]]); nAct=0; break;}
+      if (nAct==1) { merged.push_back(ori[toMerge.top()]); nAct=0; break;}
 
       auto ii = minDistToMax();      
       SingleStateVector comp;
-      comp.push_back(ori[toMerge[0]]);
+      comp.push_back(ori[toMerge.top()]);
       comp.push_back(ori[ii]);
       active[ii]=false;
-      while( (ntoMerge>0) & (!active[toMerge[0]])) {std::pop_heap(toMerge,toMerge+ntoMerge,cmp); --ntoMerge;}
+      while( (!toMerge.empty()) & (!active[toMerge.top()])) {toMerge.pop();}
       --nComp;
       nAct-=2;
       merged.push_back(MultiGaussianStateCombiner<N>().combine(comp));
-      // assert(ntoMerge>=0);
-      // assert(ntoMerge>=nAct);   
+      assert(toMerge.size()>=0);
+      assert(int(toMerge.size())>=nAct);   
     }
 
-    // std::cout << theMaxNumberOfComponents << ' ' << noComp << ' ' << nComp << ' ' << nAct << ' ' << ntoMerge << std::endl;
+    std::cout << theMaxNumberOfComponents << ' ' << noComp << ' ' << nComp << ' ' << nAct << ' ' << toMerge.size() << std::endl;
 
     if (nComp <= theMaxNumberOfComponents) { // end game
 
-        /*
+        
         {
           int kk=merged.size();
           for (int i=0; i<noComp; ++i) { if (active[i]) ++kk;}
           assert(kk==nComp);
         }
-        */
+        
  
-      MultiGaussianStateAssembler<N> result(mgs);
+      MultiGaussianStateAssembler<N> result;
 
       for (int i=0; i<noComp; ++i) { if (active[i]) result.addState(ori[i]);}
       for (auto && s : merged) result.addState(s);
 
       auto res = result.combinedState();
       
-      /*
+      
       assert(res.components().size()==oldRes.components().size());
       std::cout << res.weight() << ' ' << oldRes.weight() << std::endl;
       for(auto const & c : res.components()) std::cout << c->weight() << '/';
       std::cout << std::endl;
       for(auto const & c : oldRes.components()) std::cout << c->weight() <<	'/';
       std::cout	<< std::endl;
-      */
+      
 
       return res;
     }
@@ -188,7 +188,7 @@ CloseComponentsMerger<N>::mergeOld (const MultiState& mgs) const
     }
   }
 
-  MultiGaussianStateAssembler<N> result(mgs);
+  MultiGaussianStateAssembler<N> result;
 
   for ( typename SingleStateMap::const_iterator it = mapUnmergedComp.begin();
        it != mapUnmergedComp.end(); it++) {
@@ -216,6 +216,7 @@ CloseComponentsMerger<N>::compWithMinDistToLargestWeight(SingleStateMap& unmerge
   for (typename SingleStateMap::iterator it = unmergedComp.begin();
        it != unmergedComp.end(); it++) {
     if (it != unmergedComp.begin()) {
+      assert(unmergedComp.begin()->first >=it->first);
       double dist = (*theDistance)(*unmergedComp.begin()->second, *it->second);
       if (dist < minDist) {
 	iterMinDist = it;