--- a/TrackingTools/GsfTools/interface/CloseComponentsMerger.icc
+++ b/TrackingTools/GsfTools/interface/CloseComponentsMerger.icc
@@ -5,6 +5,9 @@
 #include <algorithm>
 #include <queue>
 #include <cfloat>
+#include "CommonTools/Utils/interface/DynArray.h"
+
+
 
 template <unsigned int N>
 CloseComponentsMerger<N>::CloseComponentsMerger (int maxNumberOfComponents,