--- a/TrackingTools/GsfTools/interface/GsfMatrixTools.h
+++ b/TrackingTools/GsfTools/interface/GsfMatrixTools.h
@@ -14,10 +14,10 @@ namespace GsfMatrixTools {
 //     return vector*matrix*vector;
 //   }
 
-  template <unsigned int N>
-  double trace (const ROOT::Math::SMatrix<double, N, N>& matrix) {
-    double result(0.);
-    for ( unsigned int i=0; i<N; i++ )  result += matrix(i,i);
+  template <typename T, unsigned int N>
+  T trace (ROOT::Math::SMatrix<T, N, N> const & matrix) {
+    T result = 0;
+    for ( unsigned int i=0; i<N; ++i )  result += matrix(i,i);
     return result;
   }
 
@@ -26,23 +26,21 @@ namespace GsfMatrixTools {
    *   a(i,j)*b(j,i) = a(i,j)*b(i,j) sum over i and j
    */
   template<typename T, unsigned int N>
-  double trace(ROOT::Math::SMatrix<T,N,N,ROOT::Math::MatRepSym<T,N> > const & a,
-	       ROOT::Math::SMatrix<T,N,N,ROOT::Math::MatRepSym<T,N> > const & b) {
-    typedef typename ROOT::Math::SMatrix<T,N,N,ROOT::Math::MatRepSym<T,N> >::const_iterator CI;
-    CI i1 = a.begin();
-    CI e1 = a.end();
-    CI i2 = b.begin();
-    //  CI e2 = b.end();
+  T trace(ROOT::Math::SMatrix<T,N,N,ROOT::Math::MatRepSym<T,N> > const & a,
+	  ROOT::Math::SMatrix<T,N,N,ROOT::Math::MatRepSym<T,N> > const & b) {
+    auto i1 = a.begin();
+    auto e1 = a.end();
+    auto i2 = b.begin();
   
     T res =0;
     // sum of the lower triangle;
     for (;i1!=e1; i1++, i2++)
       res += (*i1)*(*i2);
-    res *=2.;
+    res *= T(2.);
     // remove the duplicated diagonal...
-    for (unsigned int i=0;i<N;i++)
+    for (unsigned int i=0;i<N;++i)
       res -= a(i,i)*b(i,i);
-  return res;
+    return res;
   }
   
 }