--- a/TrackingTools/GsfTools/interface/KullbackLeiblerDistance.h
+++ b/TrackingTools/GsfTools/interface/KullbackLeiblerDistance.h
@@ -7,27 +7,18 @@
  */
 
 template <unsigned int N>
-class KullbackLeiblerDistance : public DistanceBetweenComponents<N> {
-
-private:
-  typedef typename SingleGaussianState<N>::Vector Vector;
-  typedef typename SingleGaussianState<N>::Matrix Matrix;
-  
+class KullbackLeiblerDistance final : public DistanceBetweenComponents<N> {
 public:
   
   /** Method which calculates the actual Kullback-Leibler distance.
    */
-
-  virtual double operator() (const SingleGaussianState<N>&, 
-			     const SingleGaussianState<N>&) const;
+ double operator() (const SingleGaussianState<N>&, 
+			     const SingleGaussianState<N>&) const override;
 
   virtual KullbackLeiblerDistance<N>* clone() const
   {  
     return new KullbackLeiblerDistance<N>(*this);
   }
-
-// private:
-//   double trace (const Matrix& matrix) const;
 };  
 
 #include "TrackingTools/GsfTools/interface/KullbackLeiblerDistance.icc"