--- a/TrackingTools/GsfTools/interface/KullbackLeiblerDistance.icc
+++ b/TrackingTools/GsfTools/interface/KullbackLeiblerDistance.icc
@@ -6,29 +6,22 @@ namespace KullbackLeiblerDistanceDetails {
   compute (SingleGaussianState<N>  const & sgs1, 
 	   SingleGaussianState<N>  const & sgs2) {
     
-    //ThS: That check will have to happen somewhere else for TSOS...
-  typedef ROOT::Math::SVector<double, N> Vector;
-  typedef ROOT::Math::SMatrix<double,N,N,ROOT::Math::MatRepSym<double,N> > Matrix;
+  using Vector = ROOT::Math::SVector<double, N>;
+  using Matrix = ROOT::Math::SMatrix<double,N,N,ROOT::Math::MatRepSym<double,N>>;
+
+    // a = b+c; does not vectorize for SMatrix...
      
-    //   if (&tsos1.surface() != &tsos2.surface()) {
-    //     cout << "Trying to calculate distance between components defined "
-    // 	 << "at different surfaces - returning zero!" << endl;
-    //     return 0.;
-    //   }
-    
-    const Vector& mu1 = sgs1.mean();
-    const Matrix& V1 = sgs1.covariance();
+    Vector mudiff = sgs1.mean();
+    Matrix Vdiff = sgs1.covariance();
     const Vector& mu2 = sgs2.mean();
     const Matrix& V2 = sgs2.covariance();
     
-    //   Matrix G1(V1); G1.Invert();
-    //   Matrix G2(V2); G2.Invert();
-    const Matrix& G1 = sgs1.weightMatrix();
+    Matrix Gsum = sgs1.weightMatrix();
     const Matrix& G2 = sgs2.weightMatrix();
-    Vector mudiff = mu1 - mu2;
-    Matrix Vdiff = V1 - V2;
-    Matrix Gdiff = G2 - G1;
-    Matrix Gsum = G1 + G2;
+    mudiff -= mu2;
+    Vdiff -= V2;
+    Matrix Gdiff = G2 - Gsum;
+    Gsum += G2;
     
     //   double dist = (Vdiff * Gdiff).trace() + Gsum.similarity(mudiff);
     double dist = GsfMatrixTools::trace(Vdiff,Gdiff) +
@@ -48,12 +41,3 @@ KullbackLeiblerDistance<N>::operator() (const SingleGaussianState<N> & sgs1,
   return KullbackLeiblerDistanceDetails::compute<N>(sgs1,sgs2);
   
 }
-
-// template <unsigned int N> double 
-// KullbackLeiblerDistance<N>::trace (const Matrix& matrix) const
-// {
-//   double result(0);
-//   for ( unsigned int i=0; i<N; i++ )  result += matrix(i,i);
-//   return result;
-// }
-