Major speedup of the Gsf component merging algo.
Apparently the algo (even in the original form) is not doing what was supposed to do.
This PR is purely technical. A fix of the algo logic will be submitted once fully validated.
For this reason the old algo and debug code is still left in place (commented out)

Took the opportunity to fix a long standing issue with the return of components by a TSOS
as discussed also in #13818

touches TSOS, expect recompilation of half of the universe 

negligible regressions expected due to numerics