@cmsbuild 2016-03-24
A new Pull Request was created by  @VinInn (Vincenzo Innocente) for CMSSW_8_1_X.

It involves the following packages:

CommonTools/Utils
RecoVertex/GaussianSumVertexFit
TrackingTools/GeomPropagators
TrackingTools/GsfTools
TrackingTools/GsfTracking
TrackingTools/TrajectoryState


@cmsbuild, @cvuosalo, @slava77, @davidlange6 can you please review it and eventually sign? Thanks.
@battibass, @makortel, @abbiendi, @GiacomoSguazzoni, @rovere, @VinInn, @bellan, @jhgoh, @cerati, @gpetruc, @istaslis, @trocino, @dgulhan this is something you requested to watch as well.
@slava77, @degano, @smuzaffar you are the release manager for this.

cms-bot commands are list here cms-sw/cmssw#13028