@cmsbuild 2016-03-24
The tests are being triggered in jenkins.