@cmsbuild 2016-03-24
+1
Tested at: 88894607778987133942196faf96d792b37c798f
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-13831/12056/summary.html