@cmsbuild 2016-03-25
The tests are being triggered in jenkins.