@cmsbuild 2016-03-25
+1
Tested at: 113de58d109e9c3d4cc6c412c6728838f3e4d99d
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-13831/12066/summary.html