@cvuosalo 2016-03-25
@cmsbuild please test