@cvuosalo 2016-04-01
+1

For #13831 0373457f5c755211149b455e9d0bc76bb733fc28

Speeding up GSF component merging.

The code changes are satisfactory, and Jenkins tests against baseline CMSSW_8_1_X_2016-03-31-2300 show no significant differences. Extended tests were performed on workflow 25202.0_TTbar_13 with 70 events against baseline CMSSW_8_1_X_2016-03-19-1100,  and on workflows 1317.0_SingleElectronPt35_UP15 and o1318.0_SingleGammaPt10_UP15 with 120 events each against baseline CMSSW_8_1_X_2016-03-20-0000. The results show no significant differences and no significant change in memory usage or overall CPU time, though the electronGsfTracks module is about 30% faster, so overall CPU time may be slightly reduced.

For wf 25202,the GSF-related step shows speed-up in wall-clock time:
```
  delta/mean delta/orJob     original                   new       module name
  ---------- ------------     --------                  ----       ------------
   +0.080757      +0.00%        20.50 ms/ev ->        22.22 ms/ev gedGsfElectronsTmp
   -0.252586      -0.06%       149.41 ms/ev ->       115.90 ms/ev electronGsfTracks
```
By another measure:
```
	electronGsfTracks 	 104.815 ms/ev -> 81.3096 ms/ev
```
Workflow 1317.0 (with DQM/Validation),  shows no significant change in memory use or CPU time:
```
Baseline
Max VSIZ 2955.86 on evt 10 ; max RSS 2303.42 on evt 120
Time av 1.09019 s/evt   max 20.8042 s on evt 1
M1 Time av 0.924525 s/evt   max 1.48214 s on evt 39
M8 Time av 0.924187 s/evt   max 1.48214 s on evt 39

PR 13831
Max VSIZ 2961.05 on evt 10 ; max RSS 2242.19 on evt 119
Time av 1.10107 s/evt   max 23.8653 s on evt 1
M1 Time av 0.909772 s/evt   max 1.4699 s on evt 50
M8 Time av 0.913396 s/evt   max 1.4699 s on evt 50
```
Wall-clock time for the GSF module for wf 1317.0 is shorter (excluding the first event):
```
  delta/mean delta/orJob     original                   new       module name
  ---------- ------------     --------                  ----       ------------
    -0.357626      -1.29%        36.00 ms/ev ->        25.08 ms/ev electronGsfTracks
  ---------- ------------     --------                  ----       ------------
Job total:  0.847571 s/ev ==> 0.846764 s/ev
```
By another measure:
```
	electronGsfTracks 	 47.5979 ms/ev -> 30.5983 ms/ev
	 Total in detailed printout:  	 270.395 ms/ev -> 252.442 ms/ev 	 delta: -17.953
 Total times: 443.913 ms/ev -> 428.309 ms/ev 	 delta: -15.604
```